requirements.txt
# Scraping and processing data
beautifulsoup4
country_converter
googlemaps
requests
lxml
phonenumbers

# Jupyter and related
jupyter
ipywidgets
ipyleaflet

# Google Cloud
google-cloud-firestore
google-cloud-storage

# Development needs
pylint
pytest
pytest-cov
